requirements.txt
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0

